implement FastScrollListener (fix #8406) - disable fastscroll by default (cache list, cache detail waypoint view, cache detail log view) - enable new fastscroll behavior on demand for those views - removed fastscroll for useful apps view
