+++ main/res/layout/logs_page.xml
@@ -6,7 +6,7 @@
     android:cacheColorHint="?background_color"
     android:clipToPadding="false"
     android:divider="?background_color"
-    android:fastScrollEnabled="true"
+    android:fastScrollEnabled="false"
     android:focusable="false"
     android:footerDividersEnabled="false"
     android:headerDividersEnabled="false"
