+++ main/res/layout/usefulapps_activity.xml
@@ -14,7 +14,7 @@
         android:cacheColorHint="?background_color"
         android:clipToPadding="false"
         android:divider="?background_color"
-        android:fastScrollEnabled="true"
+        android:fastScrollEnabled="false"
         android:footerDividersEnabled="false"
         android:headerDividersEnabled="false"
         android:listSelector="?background_color"
