+++ main/src/cgeo/geocaching/CacheListActivity.java
@@ -68,6 +68,7 @@ import cgeo.geocaching.sorting.CacheComparator;
 import cgeo.geocaching.sorting.SortActionProvider;
 import cgeo.geocaching.storage.DataStore;
 import cgeo.geocaching.ui.CacheListAdapter;
+import cgeo.geocaching.ui.FastScrollListener;
 import cgeo.geocaching.ui.WeakReferenceHandler;
 import cgeo.geocaching.ui.dialog.Dialogs;
 import cgeo.geocaching.utils.AndroidRxUtils;
@@ -102,7 +103,6 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AbsListView;
 import android.widget.AdapterView.AdapterContextMenuInfo;
 import android.widget.ArrayAdapter;
 import android.widget.ImageView;
@@ -185,10 +185,6 @@ public class CacheListActivity extends AbstractListActivity implements FilteredA
         }
 
     };
-    private long mLastScroll = 0; // for fast scroll control
-    private int mScrollState = 0;
-    private int mFlingStartPos = -1;
-
     private ContextMenuInfo lastMenuInfo;
     private String contextMenuGeocode = "";
     private final CompositeDisposable resumeDisposables = new CompositeDisposable();
@@ -1269,33 +1265,7 @@ public class CacheListActivity extends AbstractListActivity implements FilteredA
         adapter.setInverseSort(currentInverseSort);
         adapter.forceSort();
 
-        listView.setOnScrollListener(new AbsListView.OnScrollListener() {
-            @Override
-            public void onScrollStateChanged(final AbsListView absListView, final int state) {
-                if (state == SCROLL_STATE_IDLE && listView.isFastScrollEnabled()) {
-                    listView.postDelayed(() -> {
-                        if ((System.currentTimeMillis() - mLastScroll) > 1000) {
-                            listView.setFastScrollEnabled(false);
-                            mFlingStartPos = -1;
-                        }
-                    }, 1000);
-                }
-                mScrollState = state;
-            }
-
-            @Override
-            public void onScroll(final AbsListView view, final int firstVisibleItem, final int visibleItemCount, final int totalItemCount) {
-                mLastScroll = System.currentTimeMillis();
-                if (mScrollState == SCROLL_STATE_FLING && !listView.isFastScrollEnabled()) {
-                    if (mFlingStartPos < 0) {
-                        mFlingStartPos = firstVisibleItem;
-                    } else if (Math.abs(firstVisibleItem - mFlingStartPos) > 2) {
-                        // must have moved at least 3 entries up or down in fling mode, before fastscroll gets enabled
-                        listView.setFastScrollEnabled(true);
-                    }
-                }
-            }
-        });
+        listView.setOnScrollListener(new FastScrollListener(listView));
     }
 
     private void updateAdapter() {
